+++ src/pages/OldHome.vue
@@ -1,0 +1,16 @@
+<template>
+  <div id="app">
+    <PlioList />
+  </div>
+</template>
+
+<script>
+import PlioList from "@/components/UI/PlioList.vue";
+
+export default {
+  name: "OldHome",
+  components: {
+    PlioList,
+  },
+};
+</script>
